+++ app/(dashboard)/(routes)/teacher/analytics/_components/stripe-connect/balance-transactions.tsx
@@ -1,0 +1,86 @@
+'use client';
+
+import { format, fromUnixTime } from 'date-fns';
+
+import { getAnalytics } from '@/actions/analytics/get-analytics';
+import { TextBadge } from '@/components/common/text-badge';
+import {
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from '@/components/ui';
+import { TIMESTAMP_TEMPLATE } from '@/constants/common';
+import { DEFAULT_LOCALE } from '@/constants/locale';
+import { formatPrice, getConvertedPrice } from '@/lib/format';
+import { capitalize } from '@/lib/utils';
+
+type Analytics = Awaited<ReturnType<typeof getAnalytics>>;
+
+type BalanceTransactionsProps = { stripeConnectPayout: Analytics['stripeConnectPayouts'] };
+
+export const BalanceTransactions = ({ stripeConnectPayout }: BalanceTransactionsProps) => {
+  return (
+    <div className="w-full mt-6">
+      <Accordion type="single" collapsible>
+        <AccordionItem value="transactions" className="border-none">
+          <AccordionTrigger className="pt-0 pb-2 hover:no-underline">
+            <p>Balance Transactions</p>
+          </AccordionTrigger>
+          <AccordionContent>
+            <Table>
+              <TableHeader>
+                <TableRow>
+                  <TableHead className="w-[100px]">Type</TableHead>
+                  <TableHead>Date of transaction</TableHead>
+                  <TableHead>Status</TableHead>
+                  <TableHead>Total amount</TableHead>
+                  <TableHead>Fee amount</TableHead>
+                  <TableHead className="text-right">Net amount</TableHead>
+                </TableRow>
+              </TableHeader>
+              <TableBody>
+                {stripeConnectPayout.map((scp) => (
+                  <TableRow key={scp.id}>
+                    <TableCell className="font-medium">{capitalize(scp.type)}</TableCell>
+                    <TableCell>{format(fromUnixTime(scp.created), TIMESTAMP_TEMPLATE)}</TableCell>
+                    <TableCell>
+                      <TextBadge
+                        label={capitalize(scp.status)}
+                        variant={scp.status === 'available' ? 'green' : 'default'}
+                      />
+                    </TableCell>
+                    <TableCell>
+                      {formatPrice(getConvertedPrice(scp.amount), {
+                        locale: DEFAULT_LOCALE,
+                        currency: scp.currency,
+                      })}
+                    </TableCell>
+                    <TableCell>
+                      {formatPrice(getConvertedPrice(scp.fee), {
+                        locale: DEFAULT_LOCALE,
+                        currency: scp.currency,
+                      })}
+                    </TableCell>
+                    <TableCell className="text-right font-semibold">
+                      {formatPrice(getConvertedPrice(scp.net), {
+                        locale: DEFAULT_LOCALE,
+                        currency: scp.currency,
+                      })}
+                    </TableCell>
+                  </TableRow>
+                ))}
+              </TableBody>
+            </Table>
+          </AccordionContent>
+        </AccordionItem>
+      </Accordion>
+    </div>
+  );
+};
